feat: drop a lot of redundant files
