+++ src/models/GithubErrorResponse.ts
@@ -1,0 +1,4 @@
+export type GithubErrorResponse = {
+  message: string;
+  documentation_url?: string;
+};
