+++ src/models/GithubResponse.ts
@@ -1,0 +1,4 @@
+export type GithubResponse<T> = {
+  items: Array<T>;
+  total_count: number;
+};
