+++ src/models/GithubUser.ts
@@ -1,0 +1,3 @@
+export type GithubUser = {
+  login: string;
+};
